# src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: API School
  description: API School
  version: 1.0.0
servers:
  - url: '{url}'
    description: 'OpenAPI Documentation'
paths:
...
  /user/register:
    post:
      tags:
        - user
      operationId: registerUser
      summary: 'Create new user'
      description: 'Create new user'
      requestBody:
        description: 'Create user'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createUser'
      responses:
        201:
          description: 'User created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fullUser'
        400:
          description: 'Some properties not passed in'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        409:
          description: 'Username or email already in use'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /user/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: number
          format: int32
        required: true
    get:
      tags:
        - user
      operationId: getUserById
      summary: 'Get user'
      description: 'Get user by id'
      responses:
        200:
          description: 'User'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        404:
          description: 'User not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        400:
          description: 'Wrong parameter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      tags:
        - user
      operationId: deleteUser
      summary: 'Delete user by id'
      description: 'Delete user'
      responses:
        200:
          description: 'User deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/message'
        404:
          description: 'User not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        400:
          description: 'Wrong parameter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'

    patch:
      tags:
        - user
      operationId: updateUser
      summary: 'Get user'
      description: 'Get user by id'
      requestBody:
        description: 'Update user role'
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                role:
                  type: number
                  format: double
                  example: 2.15
      responses:
        200:
          description: 'User'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/message'
        404:
          description: 'User not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        400:
          description: 'Wrong parameter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  schemas:
...
    createUser:
      type: object
      required: username, email, password
      properties:
        username:
          type: string
          example: 'Super.Bug'
        email:
          type: string
          example: 'some@email.com'
        password:
          type: string
          example: 'password'
    fullUser:
      type: object
      properties:
        userId:
          type: number
          format: int32
          example: 12
        username:
          type: string
          example: 'Super.Bug'
        email:
          type: string
          example: 'some@email.com'
        password:
          type: string
          example: '$2a$10$kwgh9tr96dO3TaWsfhaaK.A1sj1L6MsOnUJTW41w'
        token:
          type: string
          example: 'eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjQ3LCJ'
        userRole:
          type: number
          format: int32
          example: 2
    user:
      type: object
      properties:
        id:
          type: number
          format: int32
          example: 12
        username:
          type: string
          example: 'Super.Bug'
        email:
          type: string
          example: 'some@email.com'
        role:
          type: number
          format: int32
          example: 2
...
    error:
      type: object
      properties:
        error:
          type: string
          example: 'error happens'
    message:
      type: object
      properties:
        message:
          type: string
          example: 'operation result'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT